Update exit condition for iteratespawns
The previous code was checking the same value twice, which in itself
does not even make sense. It also did not prevent the loop from looping
outside of the spawns, causing it to loop forever and teleporting
the player to the void.

The new fix makes sure that a new timer will not be created if we are
outside the current list of spawns.