Ancient spawns fix (#325)
* Remove duplicate spawn on Ancient

Spawn 16 and 17 were duplicates, which could lead into situations where
players spawn on top of each other sometimes. Removing one of them to
prevent this from happening.

* Update spawn for Ancient

Moves the player a bit out of the corner, and points them towards the
correct site. Should still provide enough cover, but should make it a
bit easier to orient when spawning.