Fix iteratespawns (#324)
* Fix usage of datapacks for iteratespawns

CreateDataTimer creates a new datapack, so if we set the values before
creating the timer, the datapack that is passed when the timer runs
would be empty. So, in a bit unintuitive way, we have to create the
timer first, and then fill in the data after.

* Update exit condition for iteratespawns

The previous code was checking the same value twice, which in itself
does not even make sense. It also did not prevent the loop from looping
outside of the spawns, causing it to loop forever and teleporting
the player to the void.

The new fix makes sure that a new timer will not be created if we are
outside the current list of spawns.